fix[PFP-92]: Adding generated bin/ directories to our .gitignore
Our internal builds generate a long long list of binary directories that we
almost certainly do not want to be part of our Github repository that
must work on many architectures and Operating Systems

However, this results in significant development pain since it's a very
easy mistake to accidentally `git add` them or even worse, miss a new file
in the very very long list.

If there is a more general way to avoid this class of error
(**/bin/ feels overzealous), please provide advice.